+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Staged
 
-* Fix for the new consent protocol [#9297](https://github.com/Applifier/unity-ads-webview/pull/9297)
+* Consent A/B test: add a new consent CTA test and remove the old one [#9271](https://github.com/Applifier/unity-ads-webview/pull/9271)
+* Deprecating ProgrammaticVastParserStrict in favor of ProgrammaticVastParser [#9298](https://github.com/Applifier/unity-ads-webview/pull/9298)
 
 # Fri Apr 26 16:24:40 UTC 2019